[FP-49] Refactor: pages 폴더, 파일명 수정 및 경로 수정
