+++ src/components/pages/activity-details/banner-image.tsx
