+++ src/pages/activities/[activity-id]/index.tsx
@@ -1,0 +1,4 @@
+const index = () => {
+  return <div></div>;
+};
+export default index;
