+++ src/pages/activities/index.tsx
@@ -1,0 +1,10 @@
+import ActivityInformation from '@/components/pages/activity-details/activity-information';
+
+const index = () => {
+  return (
+    <div>
+      <ActivityInformation />
+    </div>
+  );
+};
+export default index;
